Make the cpp *Delegator smart!
If a binder object is an input argument to a call, the delegator now
wraps that binder object in its own delegator before sending it to the
delegate.
If a binder object is output argument / return value, the delegator now
wraps the binder object that was received from the delegate with its own
delegator before return the delegator to the caller.
This way, all calls to binder objects that are sent/received through
this delegator will still go back through this delegator.

This makes it possible to have a proxy process using this delegator
between a local binder interface and a remote binder interface with
binder RPC.

Bug: 220141324
Test: atest aidl_integration_test
Change-Id: Ic83d1d28647cfd00161b29e3da3ff21a7065c29e